post

+++ DegreeProject/DegreeProject.DTO/Projects/StandartDTO.cs
@@ -13,7 +13,7 @@ namespace DegreeProject.DTO.Projects
         public string NameResourse { get; set; }
         public string Unit { get; set; }
         public int UnitAmount { get; set; }
-        public decimal LaborCostHour { get; set; }
-        public decimal LaborCostMachine { get; set; }
+        public double LaborCostHour { get; set; }
+        public double LaborCostMachine { get; set; }
     }
 }
